requirements.txt
fastapi==0.115.5
langchain_core==0.3.18
langchain_openai==0.2.8
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.32.0
